Init
halp me

+++ src/Calculator.java
@@ -2,16 +2,16 @@ import java.util.Scanner;
 public class Calculator {
 	private static Scanner in;
 	public static void main(String[] args) {
-		float a;
-		float b;
-		float c;
+		double a;
+		double b;
+		double c;
 		in = new Scanner(System.in);
 		System.out.println("Enter an integer");  
-		 a = in.nextFloat();
+		 a = in.nextDouble();
 		 System.out.println("Enter another integer");
-		 b = in.nextFloat();
+		 b = in.nextDouble();
 		 System.out.println("Press1 for +, 2 for -, 3 for * or 4 for /");
-		 c = in.nextFloat();
+		 c = in.nextDouble();
 		 if(c == 1){
 			 System.out.println(a + b);}
 		if(c == 2){
@@ -20,7 +20,7 @@ public class Calculator {
 			System.out.println(a*b);
 		}
 		if(c == 4){
-			float g = a/b;
+			double g = a/b;
 			System.out.println(g);
 		}
